module: DOM types
