+++ types/node/ts5.7/index.d.ts
@@ -47,6 +47,7 @@
 /// <reference path="../web-globals/domexception.d.ts" />
 /// <reference path="../web-globals/events.d.ts" />
 /// <reference path="../web-globals/fetch.d.ts" />
+/// <reference path="../web-globals/importmeta.d.ts" />
 /// <reference path="../web-globals/navigator.d.ts" />
 /// <reference path="../web-globals/storage.d.ts" />
 /// <reference path="../web-globals/streams.d.ts" />
