+++ types/node/web-globals/importmeta.d.ts
@@ -1,0 +1,13 @@
+export {};
+
+import { URL } from "node:url";
+
+declare global {
+    interface ImportMeta {
+        dirname: string;
+        filename: string;
+        main: boolean;
+        url: string;
+        resolve(specifier: string, parent?: string | URL): string;
+    }
+}
